fix monkey-patching
